ci-based/requirements.txt
Flask
requests
cryptography
six
pyyaml
rq==1.15.1
